Created TmuxRunner
With TmuxRunner you can register "frameworks" to run tests, or other
commands in tmux validating if the current file is runnable.

A TmuxRunner framework should implements a validator and a runner, the
former receive the current file and checks if it's runnable. The latter
receive the current file and a parameter if the framework should run
'all' files, current 'file' or 'current' context.