Add closer.
Closer is a plugin(draft) to help get out from pair closers.
To open a pair use `sym<tab>`, write the content the next <tab> will get
out from the closer sym, except the content is a snippet, in this case
the snippet will be expanded. Example(the _ is the cursor position):

without snippet:
"<tab>                   # open a pair with "
"_"                      # result after first <tab>
"test<tab>"_             # after the content, with the second <tab>
                         # the cursor will be positioned after the
                         # second "

with snippet:
"<tab>                   # open a pair with "
"_"                      # result after first <tab>
"func<tab>"              # after the content, with the second <tab>
"function! _(<2>) ...    # if the content is a snippet, this will be
                         # expanded