Avoid signify to override neomake signs
I don't know where, but somewhere on my configuration has an
`autocmd FileType javascript let g:neomake_place_signs = 0`
which makes the signify signs override the neomake ones.  I searched in
a lot of codes, tried a lot of custom vimrc files, and wasn't able to
found the where the autocmd is defined.

The solution I found, which is not good, was to add a `dictwatcheradd`
on the `g:neomake_place_signs` to enforce it always to be `1`.